+++ src/main/java/com/lawencon/payroll/config/SecurityConfig.java
@@ -42,7 +42,7 @@ public class SecurityConfig {
 	@Bean
 	public List<RequestMatcher> matchers() {
 		final List<RequestMatcher> matchers = new ArrayList<>();
-		matchers.add(new AntPathRequestMatcher("/users/login", HttpMethod.POST.name()));
+		matchers.add(new AntPathRequestMatcher("/login/", HttpMethod.POST.name()));
 		matchers.add(new AntPathRequestMatcher("/files/**", HttpMethod.GET.name()));
 		matchers.add(new AntPathRequestMatcher("/documents/download/**", HttpMethod.GET.name()));
 		matchers.add(new AntPathRequestMatcher("/documents/preview/**", HttpMethod.GET.name()));
