+++ src/main/java/com/lawencon/payroll/service/LoginService.java
@@ -1,0 +1,8 @@
+package com.lawencon.payroll.service;
+
+import com.lawencon.payroll.dto.user.LoginReqDto;
+import com.lawencon.payroll.dto.user.LoginResDto;
+
+public interface LoginService {
+    LoginResDto loginUser(LoginReqDto data);
+}
